Chained ctor of AggregateType
