+++ src/tests/Microsoft.PowerFx.Core.Tests/DisplayNameTests.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Globalization;
 using Microsoft.PowerFx.Core;
 using Microsoft.PowerFx.Core.Binding;
@@ -18,6 +20,80 @@ namespace Microsoft.PowerFx.Interpreter.Tests
 {
     public class DisplayNameTests : PowerFxTest
     {
+        public class LazyRecordType : RecordType
+        {
+            public override IEnumerable<string> FieldNames { get; }
+
+            public override bool TryGetFieldType(string name, out FormulaType type)
+            {
+                type = name switch
+                {
+                    "Num" => FormulaType.Number,
+                    "B" => FormulaType.Boolean,
+                    "Nested" => FormulaType.Unknown,
+                    "Inner" => FormulaType.Number,
+                    _ => FormulaType.Blank
+                };
+
+                return type != FormulaType.Blank;
+            }
+
+            public LazyRecordType()
+                : base(new CustomDisplayNameProvider())
+            {
+                FieldNames = new List<string>() { "Num", "B", "Nested", "Inner" };
+            }
+
+            public override bool Equals(object other)
+            {
+                return other is LazyRecordType;
+            }
+
+            public override int GetHashCode()
+            {
+                return 3;
+            }
+
+            private class CustomDisplayNameProvider : DisplayNameProvider
+            {
+                internal override ImmutableDictionary<DName, DName> LogicalToDisplayPairs => throw new NotImplementedException();
+
+                internal override bool TryGetDisplayName(DName logicalName, out DName displayDName)
+                {
+                    var displayName = logicalName.Value switch
+                    {
+                        "Num" => "DisplayNum",
+                        "B" => "DisplayB",
+                        "Inner" => "InnerDisplay",
+                        "Nested" => "NestedDisplay",
+                        _ => null
+                    };
+                    displayDName = displayName == null ? default : new DName(displayName);
+                    return displayName != null;
+                }
+
+                internal override bool TryGetLogicalName(DName displayName, out DName logicalDName)
+                {
+                    var logicalName = displayName.Value switch
+                    {
+                        "DisplayNum" => "Num",
+                        "DisplayB" => "B",
+                        "InnerDisplay" => "Inner",
+                        "NestedDisplay" => "Nested",
+                        _ => null
+                    };
+                    logicalDName = logicalName == null ? default : new DName(logicalName);
+                    return logicalName != null;
+                }
+
+                internal override bool TryRemapLogicalAndDisplayNames(DName displayName, out DName logicalName, out DName newDisplayName)
+                {
+                    newDisplayName = displayName;
+                    return TryGetLogicalName(displayName, out logicalName);
+                }
+            }
+        }
+
         public DisplayNameTests()
             : base()
         {
@@ -88,27 +164,23 @@ namespace Microsoft.PowerFx.Interpreter.Tests
             var result = _engine.Check(inputExpression, r1);
             Assert.True(result.IsSuccess);
 
-            if (toDisplay)
-            {
-                var outDisplayExpression = _engine.GetDisplayExpression(inputExpression, r1);
-                Assert.Equal(outputExpression, outDisplayExpression);
-            }
-            else
-            {
-                var outInvariantExpression = _engine.GetInvariantExpression(inputExpression, r1);
-                Assert.Equal(outputExpression, outInvariantExpression);
-            }
+            // Below Record r2 Tests the second method where we provide DisplayNameProvider via constructor to 
+            // initialize the DisplayNameProvider for derived record types.
+            var r2 = new LazyRecordType();
 
-            r1 = new LazyRecordType();
-            if (toDisplay)
+            var records = new RecordType[] { r1, r2 };
+            foreach (var r in records)
             {
-                var outDisplayExpression = _engine.GetDisplayExpression(inputExpression, r1);
-                Assert.Equal(outputExpression, outDisplayExpression);
-            }
-            else
-            {
-                var outInvariantExpression = _engine.GetInvariantExpression(inputExpression, r1);
-                Assert.Equal(outputExpression, outInvariantExpression);
+                if (toDisplay)
+                {
+                    var outDisplayExpression = _engine.GetDisplayExpression(inputExpression, r1);
+                    Assert.Equal(outputExpression, outDisplayExpression);
+                }
+                else
+                {
+                    var outInvariantExpression = _engine.GetInvariantExpression(inputExpression, r1);
+                    Assert.Equal(outputExpression, outInvariantExpression);
+                }
             }
         }
