--- src/tests/Microsoft.PowerFx.Core.Tests/LazyRecordType.cs
@@ -1,89 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Globalization;
-using System.Text;
-using Microsoft.PowerFx.Core.Types;
-using Microsoft.PowerFx.Core.Utils;
-using Microsoft.PowerFx.Types;
-using Xunit;
-
-namespace Microsoft.PowerFx.Core.Tests
-{
-    public class LazyRecordType : RecordType
-    {
-        public override IEnumerable<string> FieldNames { get; }
-
-        public override bool TryGetFieldType(string name, out FormulaType type)
-        {
-            type = name switch
-            {
-                "Num" => FormulaType.Number,
-                "B" => FormulaType.Boolean,
-                "Nested" => FormulaType.Unknown,
-                "Inner" => FormulaType.Number,
-                _ => FormulaType.Blank
-            };
-
-            return type != FormulaType.Blank;
-        }
-
-        public LazyRecordType()
-            : base(new CustomDisplayNameProvider())
-        {
-            FieldNames = new List<string>() { "Num", "B", "Nested", "Inner" };
-        }
-
-        public override bool Equals(object other)
-        {
-            return other is LazyRecordType;
-        }
-
-        public override int GetHashCode()
-        {
-            return 3;
-        }
-
-        private class CustomDisplayNameProvider : DisplayNameProvider
-        {
-            internal override ImmutableDictionary<DName, DName> LogicalToDisplayPairs => throw new NotImplementedException();
-
-            internal override bool TryGetDisplayName(DName logicalName, out DName displayDName)
-            {
-                var displayName = logicalName.Value switch
-                {
-                    "Num" => "DisplayNum",
-                    "B" => "DisplayB",
-                    "Inner" => "InnerDisplay",
-                    "Nested" => "NestedDisplay",
-                    _ => null
-                };
-                displayDName = displayName == null ? default : new DName(displayName);
-                return displayName != null;
-            }
-
-            internal override bool TryGetLogicalName(DName displayName, out DName logicalDName)
-            {
-                var logicalName = displayName.Value switch
-                {
-                    "DisplayNum" => "Num",
-                    "DisplayB" => "B",
-                    "InnerDisplay" => "Inner",
-                    "NestedDisplay" => "Nested",
-                    _ => null
-                };
-                logicalDName = logicalName == null ? default : new DName(logicalName);
-                return logicalName != null;
-            }
-
-            internal override bool TryRemapLogicalAndDisplayNames(DName displayName, out DName logicalName, out DName newDisplayName)
-            {
-                newDisplayName = displayName;
-                return TryGetLogicalName(displayName, out logicalName);
-            }
-        }
-    }
-}
